+++ providers/bing/index.ts
@@ -1,0 +1,2 @@
+import * as bing from './bing'
+export = bing