+++ providers/google/index.ts
@@ -1,0 +1,2 @@
+import * as google from './google'
+export = google