+++ providers/mapbox/index.ts
@@ -1,0 +1,2 @@
+import * as mapbox from './mapbox'
+export = mapbox