+++ providers/wikidata/index.ts
@@ -1,0 +1,2 @@
+import * as wikidata from './wikidata'
+export = wikidata